requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.10.0
graphene==3.2.2
graphene-django==3.0.2
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
